PR fix2

+++ parcel.go
@@ -82,21 +82,35 @@ func (s ParcelStore) SetStatus(number int, status string) error {
 func (s ParcelStore) SetAddress(number int, address string) error {
 	// реализуйте обновление адреса в таблице parcel
 	// менять адрес можно только если значение статуса registered
-	_, err := s.db.Exec("UPDATE parcel SET address = :address WHERE number = :number",
-		sql.Named("address", number),
-		sql.Named("number", number))
+	p, err := s.Get(number)
 	if err != nil {
 		return fmt.Errorf("ParcelStore SetAddress error: %w", err)
 	}
-	return nil
+	if p.Status == ParcelStatusRegistered {
+		_, err = s.db.Exec("UPDATE parcel SET address = :address WHERE number = :number",
+			sql.Named("address", number),
+			sql.Named("number", number))
+		if err != nil {
+			return fmt.Errorf("ParcelStore SetAddress error: %w", err)
+		}
+		return nil
+	}
+	return fmt.Errorf("ParcelStore SetAddress error: bad status - %s", p.Status)
 }
 
 func (s ParcelStore) Delete(number int) error {
 	// реализуйте удаление строки из таблицы parcel
 	// удалять строку можно только если значение статуса registered
-	_, err := s.db.Exec("DELETE FROM parcel WHERE number = :number", sql.Named("number", number))
+	p, err := s.Get(number)
 	if err != nil {
-		return fmt.Errorf("ParcelStore Delete error: %w", err)
+		return fmt.Errorf("ParcelStore SetAddress error: %w", err)
 	}
-	return nil
+	if p.Status == ParcelStatusRegistered {
+		_, err := s.db.Exec("DELETE FROM parcel WHERE number = :number", sql.Named("number", number))
+		if err != nil {
+			return fmt.Errorf("ParcelStore Delete error: %w", err)
+		}
+		return nil
+	}
+	return fmt.Errorf("ParcelStore Delete error: bad status - %s", p.Status)
 }
